+++ app/src/debug/java/org/simple/clinic/di/network/HttpInterceptorsModule.kt
@@ -1,5 +1,7 @@
 package org.simple.clinic.di.network
 
+import android.app.Application
+import com.chuckerteam.chucker.api.ChuckerInterceptor
 import dagger.Module
 import dagger.Provides
 import io.sentry.okhttp.SentryOkHttpInterceptor
@@ -12,10 +14,16 @@ import org.simple.clinic.user.LoggedInUserHttpInterceptor
 class HttpInterceptorsModule {
 
   @Provides
+  fun chuckerInterceptor(context: Application): ChuckerInterceptor {
+    return ChuckerInterceptor(context)
+  }
+
+  @Provides
   fun providerInterceptors(
       loggedInInterceptor: LoggedInUserHttpInterceptor,
       appInfoHttpInterceptor: AppInfoHttpInterceptor,
-      compressRequestInterceptor: CompressRequestInterceptor
+      compressRequestInterceptor: CompressRequestInterceptor,
+      chuckerInterceptor: ChuckerInterceptor,
   ): List<Interceptor> {
     val loggingInterceptor = HttpLoggingInterceptor().apply {
       level = BODY
@@ -26,7 +34,8 @@ class HttpInterceptorsModule {
         loggedInInterceptor,
         appInfoHttpInterceptor,
         loggingInterceptor,
-        compressRequestInterceptor
+        compressRequestInterceptor,
+        chuckerInterceptor,
     )
   }
 }